refactor: centralize keyword consumption logic
