+++ src/lexer/ImportMetaReader.js
@@ -1,23 +1,8 @@
+import { consumeKeyword } from './utils.js';
+
 export function ImportMetaReader(stream, factory) {
   const startPos = stream.getPosition();
-  const seq = 'import.meta';
-  if (!stream.input.startsWith(seq, startPos.index)) return null;
-
-  const saved = stream.getPosition();
-  for (const ch of seq) {
-    if (stream.current() !== ch) {
-      stream.setPosition(saved);
-      return null;
-    }
-    stream.advance();
-  }
-
-  const next = stream.current();
-  if (next && /[A-Za-z0-9_$]/.test(next)) {
-    stream.setPosition(saved);
-    return null;
-  }
-
-  const endPos = stream.getPosition();
-  return factory('IMPORT_META', seq, startPos, endPos);
+  const endPos = consumeKeyword(stream, 'import.meta', { checkPrev: false });
+  if (!endPos) return null;
+  return factory('IMPORT_META', 'import.meta', startPos, endPos);
 }
